requirements.txt
streamlit==0.84.0
Pillow==8.0.1
numpy==1.19.4
pypng==0.0.20
PyQRCode==1.2.1
base58==2.1.0